Moved personal-detail component into enrolment - builds & serves
