+++ src/app/modules/enrolment/components/personal-details/personal-details.component.ts
@@ -2,16 +2,9 @@ import {
   Component,
   Input,
   Output,
-  OnChanges,
   EventEmitter,
-  SimpleChange,
   ViewChild,
-  AfterViewInit,
-  OnInit,
-  OnDestroy,
   ElementRef,
-  QueryList,
-  ViewChildren,
   ChangeDetectorRef
 } from '@angular/core';
 import { state, trigger, style } from '@angular/animations';
@@ -33,28 +26,26 @@ import {
 import { MspImage } from '../../../../models/msp-image';
 import * as _ from 'lodash';
 
-import { MspIdReqModalComponent } from '../id-req-modal/id-req-modal.component';
-import { MspImageErrorModalComponent } from '../image-error-modal/image-error-modal.component';
-import { MspBirthDateComponent } from '../birthdate/birthdate.component';
 import { MspGenderComponent } from '../../../../components/msp/common/gender/gender.component';
 import { MspSchoolDateComponent } from '../../../../components/msp/common/schoolDate/school-date.component';
 import { HealthNumberComponent } from '../../../../components/msp/common/health-number/health-number.component';
 import { MspDischargeDateComponent } from '../../../../components/msp/common/discharge-date/discharge-date.component';
-import { MspAddressComponent } from '../address/address.component';
 
 import { MspArrivalDateComponent } from '../../../../components/msp/common/arrival-date/arrival-date.component';
 import { MspOutofBCRecordComponent } from '../../../../components/msp/common/outof-bc/outof-bc.component';
 import { BaseComponent } from '../../../../models/base.component';
-import { ServicesCardDisclaimerModalComponent } from '../services-card-disclaimer/services-card-disclaimer.component';
 import {
   CANADA,
   Address,
-  ProvinceList,
-  BRITISH_COLUMBIA
-} from 'moh-common-lib';
+  ProvinceList} from 'moh-common-lib';
 import { MspAddressConstants } from '../../../../models/msp-address.constants';
 import { MspDocumentConstants } from '../../../../models/msp-document.constants';
 import { legalStatus } from '../../../../models/msp.contants';
+import { MspIdReqModalComponent } from '../../../msp-core/components/id-req-modal/id-req-modal.component';
+import { MspImageErrorModalComponent } from '../../../msp-core/components/image-error-modal/image-error-modal.component';
+import { MspBirthDateComponent } from '../../../msp-core/components/birthdate/birthdate.component';
+import { MspAddressComponent } from '../../../msp-core/components/address/address.component';
+import { ServicesCardDisclaimerModalComponent } from '../../../msp-core/components/services-card-disclaimer/services-card-disclaimer.component';
 
 @Component({
   selector: 'msp-personal-details',
