+++ src/app/modules/msp-core/msp-core.module.ts
@@ -4,7 +4,6 @@ import { MspFullNameComponent } from './components/full-name/full-name.component
 import { SharedCoreModule } from 'moh-common-lib';
 import { CaptchaModule } from 'moh-common-lib/captcha';
 import { FormsModule } from '@angular/forms';
-import { PersonalDetailsComponent } from './components/personal-details/personal-details.component';
 import { CitizenStatusComponent } from './components/citizen-status/citizen-status.component';
 import { ServicesCardDisclaimerModalComponent } from './components/services-card-disclaimer/services-card-disclaimer.component';
 import { MspStatusInCanadaRadioComponent } from './components/status-in-canada-radio/status-in-canada-radio.component';
@@ -58,7 +57,6 @@ import { AddressRetroSuppbenComponent } from './components/address-retro-suppben
 
 const componentList = [
   MspFullNameComponent,
-  PersonalDetailsComponent,
   CitizenStatusComponent,
   ServicesCardDisclaimerModalComponent,
   MspStatusInCanadaRadioComponent,
